Update to .NET 9
- Upgraded target framework from `net8.0` to `net9.0` and language version from `12` to `13`.
- Refactored connection handling and error management in `MainWindow`.
- Changed from `"1"` to `"en"` for string retrieval.
- Enhanced `GenerateWebhook` method to include shiny criteria logic.
- Updated memory addresses in `Offsets.cs` for raid block pointers.
- Altered `IsNatureSatisfied` method in `RaidFilter` from `int` to `byte` and optimized `IsIVsSatisfied` method.
- Refactored package references in `RaidCrawler.Core.csproj` for centralized management in `Directory.Packages.props`.
- Added `Directory.Packages.props` for consistent package versioning across the project.

Big thanks to @santacrab2 and @Manu098vm for the help learning how to debug and fix these issues.

+++ RaidCrawler.Tests/RaidStatTests.cs
@@ -36,7 +36,7 @@ public class RaidStatTests : TestUtil
         Encounter9RNG.GenerateData(blank, param, EncounterCriteria.Unrestricted, raid.Seed);
         var encIVs = Utils.ToSpeedLast(blank.IVs);
         encIVs.SequenceEqual(ivs).Should().BeTrue();
-        blank.Nature.Should().Be((int)nature);
+        blank.Nature.Should().Be(nature);
         blank.IsShiny.Should().Be(shiny);
     }
 }
